Modernize ProcessingTimelineComponent component
- Use typescript and type definitions
- Replace `createReactClass` with functional component
- Fix linter errors
